+++ healthper/src/main/java/umc/healthper/repository/post/PostRepositoryCustom.java
@@ -1,4 +1,4 @@
-package umc.healthper.repository;
+package umc.healthper.repository.post;
 
 public interface PostRepositoryCustom {
     void removePost(Long id);
