+++ healthper/src/main/java/umc/healthper/service/PostService.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.healthper.domain.Post;
 import umc.healthper.domain.PostStatus;
 import umc.healthper.dto.post.UpdatePostDto;
-import umc.healthper.repository.PostRepository;
+import umc.healthper.repository.post.PostRepository;
 
 @Service
 @Transactional(readOnly = true)
